Se acomodan archivos, aprovado por Diego. Gabriel, Edgar e Ignacio
