+++ src/funcionesGPS.py
@@ -1,14 +1,6 @@
 import serial
 import requests
 import csv
-import os
-import time
-
-# Se detecta si el os es windows para simular
-# la toma de datos (solo para pruebas fuera de la Raspberry)
-windows = False
-if os.name == 'nt':
-    windows = True
 
 
 def guardar_csv(nombre_archivo, latitud, longitud):
@@ -97,18 +89,12 @@ def enviar_api(latitud_decimal, longitud_decimal):
     """
     # Datos que se enviarán a la API
     data = {
-        "journey_id": 698453,
-        "timestamp": 1710067980,
         "latitude": latitud_decimal,
         "longitude": longitud_decimal,
-        "altitude": 1234.5,
-        "speed": 12.5,
-        "bearing": 135,
-        "odometer": 12345.6
     }
 
     # URL de la API
-    api_url = 'https://realtime.bucr.digital/api/position'
+    api_url = "https://api-tcu-ucr-default-rtdb.firebaseio.com/location.json"
 
     try:
         # Enviar datos a la API
@@ -142,8 +128,8 @@ def manejarGPS(stop_event):
     # api_url = 'https://realtime.bucr.digital/api/position'
 
     # Se abre puerto serial
-    if not windows:
-        ser = serial.Serial(port, baudrate, timeout=10)
+    ser = serial.Serial(port, baudrate, timeout=10)
+
     # Contador de guardados
     save_count = 1
 
@@ -152,59 +138,40 @@ def manejarGPS(stop_event):
             f"Guardando datos en {archivo_csv} y "
             f"{archivo_txt}. Presione CTRL+C para salir.")
         while not stop_event.is_set():
-            if windows:
-                latitud = 1
-                longitud = 2
-                print("Proceso corriendo...")
-                time.sleep(0.5)
-                with open(archivo_txt, 'a') as txt_file:
-                    # Guarda la línea en el archivo de texto
-                    txt_file.write(
-                        f"{save_count}: Latitud: {latitud},"
-                        f" Longitud: {longitud} \n")
+            line = ser.readline().decode('ascii', errors='replace').strip(
+            )  # Lee datos del puerto serial
+            # Verifica si la línea comienza con $GPRMC
+            if line.startswith('$GPRMC'):
+                # Genera una lista de 13 elementos
+                lista = line.split(',')
+                # Si la lista contiene mas de 2 elementos
+                if len(lista) > 2:
+                    # Si el elemento 3 es A (dato valido)
+                    if lista[2] == 'A':
+                        # Guardar elementos 4 al 8
+                        # [latitud,latitud_dir,longitud,longitud_dir]
+                        latxlon = lista[3:7]
+                        latitud, longitud = conversion_latxlon(
+                            latxlon[0], latxlon[1], latxlon[2], latxlon[3])
+                        print(f"Latitud: {latitud} y Longitud: {longitud}")
 
-                guardar_csv(archivo_csv, latitud, longitud)
-                guardar_txt(archivo_txt, latitud, longitud,
-                            save_count)
-                # Incrementa el contador de guardados
-                save_count += 1
-            else:
-                line = ser.readline().decode('ascii', errors='replace').strip(
-                )  # Lee datos del puerto serial
-                # Verifica si la línea comienza con $GPRMC
-                if line.startswith('$GPRMC'):
-                    # Genera una lista de 13 elementos
-                    lista = line.split(',')
-                    # Si la lista contiene mas de 2 elementos
-                    if len(lista) > 2:
-                        # Si el elemento 3 es A (dato valido)
-                        if lista[2] == 'A':
-                            # Guardar elementos 4 al 8
-                            # [latitud,latitud_dir,longitud,longitud_dir]
-                            latxlon = lista[3:7]
-                            latitud, longitud = conversion_latxlon(
-                                latxlon[0], latxlon[1], latxlon[2], latxlon[3])
-                            print(f"Latitud: {latitud} y Longitud: {longitud}")
+                        with open(archivo_txt, 'a') as txt_file:
+                            # Guarda la línea en el archivo de texto
+                            txt_file.write(
+                                f"{save_count}: Latitud: {latitud},"
+                                f" Longitud: {longitud} \n")
 
-                            with open(archivo_txt, 'a') as txt_file:
-                                # Guarda la línea en el archivo de texto
-                                txt_file.write(
-                                    f"{save_count}: Latitud: {latitud},"
-                                    f" Longitud: {longitud} \n")
+                        enviar_api(latitud, longitud)
+                        guardar_csv(archivo_csv, latitud, longitud)
+                        guardar_txt(archivo_txt, latitud, longitud,
+                                    save_count)
+                        # Incrementa el contador de guardados
+                        save_count += 1
 
-                            guardar_csv(archivo_csv, latitud, longitud)
-                            guardar_txt(archivo_txt, latitud, longitud,
-                                        save_count)
-                            # Incrementa el contador de guardados
-                            save_count += 1
-
-                        elif lista[2] == 'V':
-                            print(
-                                f"Linea completa: {line}\nLista:"
-                                f" None\nLínea inválida (V)")
+                    elif lista[2] == 'V':
+                        print(
+                            f"Linea completa: {line}\nLista:"
+                            f" None\nLínea inválida (V)")
 
     except KeyboardInterrupt:
         print("\nPrograma interrumpido. Cerrando...")
-    finally:
-        if not windows:
-            ser.close()
